ci: upgrade dotnet version from 3 to 8
This primarily fixes the issue where "no usable version of libssl was found" (see [this post](https://stackoverflow.com/a/72836925)) as newer Ubuntu versions upgraded OpenSSL to version 3.0, which are not supported by dotnet versions 5 and older. Therefore, we can upgrade the local version of dotnet to one that does support this version of OpenSSL. This has the added benefit of using a more up-to-date LTS version of dotnet as v3.1 reached EOL at the end of 2022.